SMGCPA: Fix wrong handling of edges overlap on memcpy [Merged from klever_fixes revision 36838]
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38368 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/evaluator/ExpressionValueVisitor.java
@@ -260,7 +260,7 @@ class ExpressionValueVisitor
         return result.build();
 
     case SIZEOF:
-        int size =
+        long size =
             smgExpressionEvaluator.getBitSizeof(
                 cfaEdge,
                 TypeUtils.getRealExpressionType(unaryOperand),
